# ai-agent-fe/openapi.yaml
openapi: 3.1.0
info:
  title: AI Agent API
  description: API for the AI Agent frontend and service interaction
  version: 0.1.0
servers:
  - url: /api
    description: Development server

# --- Components Definition (Single top-level key) ---
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error
      required:
        - error

    # --- User Schemas ---
    UserProfilePartial:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        planTier:
          type: string
          enum: [Free, Pro, Enterprise]
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        settings:
          type: object
          additionalProperties: true
          nullable: true
        usageCounters: # Optional, depends on implementation
          type: object
          properties:
            month:
              type: string
            messagesUsed:
              type: integer
            projectsGenerated:
              type: integer
            tokensUsed:
              type: integer
              nullable: true
          nullable: true
    UserUpdateInput:
      type: object
      properties:
        name:
          type: string
        settings:
          type: object
          additionalProperties: true
      minProperties: 1 # Must provide at least one field to update
    UserUsageResponse:
      type: object
      properties:
        planTier:
          type: string
          enum: [Free, Pro, Enterprise]
        messagesUsed:
          type: integer
        messagesLimit:
          type: number # Can be Infinity
        projectsGenerated:
          type: integer
        projectsLimit:
          type: number # Can be Infinity
        tokensUsed:
          type: integer
          nullable: true
        tokensLimit:
          type: number # Can be Infinity

    # --- Session Schemas ---
...
    ProjectFile:
        type: object
        properties:
            path:
                type: string
            content:
                type: string
                nullable: true
            size:
                type: integer
                nullable: true
            storageRef:
                type: string
                nullable: true
    ProjectSummaryResponse:
      type: object
      properties:
        projectId:
          type: string
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        previewURL:
          type: string
          format: url
          nullable: true
        sessionId:
          type: string
          nullable: true
    ProjectDetailResponse:
        allOf:
          - $ref: '#/components/schemas/ProjectSummaryResponse'
          - type: object
            properties:
              storagePath:
                type: string
                nullable: true
              boltContainerId:
                type: string
                nullable: true
              collaboratorIds:
                type: array
                items:
                  type: string
                  format: uuid
                nullable: true
              isPublic:
                type: boolean
                nullable: true
              files:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectFile'
                nullable: true
    ProjectUpdateInput:
        type: object
        properties:
            name:
                type: string
            description:
                type: string
        minProperties: 1
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePartial'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateInput'
      responses:
        '200':
          description: User profile updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePartial' # Returns partial updated profile
        '400':
          description: Bad Request (e.g., no fields provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me/usage:
    get:
      tags: [Users]
      summary: Get current user usage statistics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usage statistics retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUsageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Usage data not found for user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
        description: ID of the project.
    get:
      tags: [Projects]
      summary: Get project details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Projects]
      summary: Update project metadata
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateInput'
      responses:
        '200':
          description: Project updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse' # Returns partial updated project
        '400':
          description: Bad Request (e.g., no fields provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Projects]
      summary: Delete a project
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Project deleted successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...